Labfiles/content-understanding/python/requirements.txt
python-dotenv
azure-identity
requests